requirements.txt
opencv-python
pillow
tensorflow
matplotlib
numpy
pandas
scikit-learn